+++ framework/CHANGELOG.md
@@ -9,12 +9,28 @@ Yii Framework 2 Change Log
 - Bug #18026: Fix `ArrayHelper::getValue()` did not work with `ArrayAccess` objects (mikk150)
 - Enh #18048: Use `Instance::ensure()` to set `User::$accessChecker` (lav45)
 - Bug #18051: Fix missing support for custom validation method in EachValidator (bizley)
+- Enh #17722: Add action injection support (SamMousa, samdark)
 - Bug #18041: Fix RBAC migration for MSSQL (darkdef)
 - Bug #18081: Fix for PDO_DBLIB/MSSQL. Set flag ANSI_NULL_DFLT_ON to ON for current connect to DB (darkdef)
 - Bug #13828: Fix retrieving inserted data for a primary key of type uniqueidentifier for SQL Server 2005 or later (darkdef)
 - Bug #17474: Fix retrieving inserted data for a primary key of type trigger for SQL Server 2005 or later (darkdef)
 - Bug #18001: Fix getting table metadata for tables `(` in their name (floor12)
 - Bug #18096: Fix InlineValidator with anonymous inline function not working well from EachValidator (trombipeti)
+- Enh #18102: Use “primary”/“replica” terminology instead of “master”/“slave” (brandonkelly)
+  - Added `yii\db\Connection::$enableReplicas` and deprecated `$enableSlaves` via magic methods.
+  - Added `yii\db\Connection::$replicas` and deprecated `$slaves` via magic methods.
+  - Added `yii\db\Connection::$replicaConfig` and deprecated `$slaveConfig` via magic methods.
+  - Added `yii\db\Connection::$primaries` and deprecated `$masters` via magic methods.
+  - Added `yii\db\Connection::$primaryConfig` and deprecated `$masterConfig` via magic methods.
+  - Added `yii\db\Connection::$shufflePrimaries` and deprecated `$shuffleMasters` via magic methods.
+  - Added `yii\db\Connection::getReplicaPdo()` and deprecated `getSlavePdo()`.
+  - Added `yii\db\Connection::getPrimaryPdo()` and deprecated `getMasterPdo()`.
+  - Added `yii\db\Connection::getReplica()` and deprecated `getSlave()`.
+  - Added `yii\db\Connection::getPrimary()` and deprecated `getMaster()`.
+  - Added `yii\db\Connection::usePrimary()` and deprecated `useMaster()`.
+  - Added `yii\validators\ExistValidator::$forcePrimaryDb` and deprecated `$forceMasterDb` via magic methods.
+  - Added `yii\validators\UniqueValidator::$forcePrimaryDb` and deprecated `$forceMasterDb` via magic methods.
+- Bug #18101: Fix behavior of OUTPUT INSERTED.* for SQL Server query: "insert default values"; correct MSSQL unit tests; turn off profiling echo message in migration test (darkdef)
 
 
 2.0.35 May 02, 2020